feat: Add Database Routing for v1 Analytics Data API URLs
The new v1 URLs will be supported by the new Aurora MySQL database. In order to ensure that v1 views are backed by this database, this pull request adds custom middleware to add a database "hint" to thread-local storage when handling v1 URLs. This "hint" is used by the AnalyticsApiRouter database router to route model operations to the Aurora MySQL database instead of the original database.

JIRA: https://openedx.atlassian.net/browse/MST-1095
