+++ analytics_data_api/v0/tests/views/test_course_summaries.py
@@ -9,7 +9,7 @@ from django_dynamic_fixture import G
 from analytics_data_api.constants import enrollment_modes
 from analytics_data_api.v0 import models, serializers
 from analytics_data_api.v0.tests.views import APIListViewTestMixin, CourseSamples, VerifyCourseIdMixin
-from analyticsdataserver.tests import TestCaseWithAuthentication
+from analyticsdataserver.tests.utils import TestCaseWithAuthentication
 
 
 @ddt.ddt
