+++ analytics_data_api/v0/tests/views/test_enterprise_learner_engagements.py
@@ -13,7 +13,7 @@ from analytics_data_api.constants.engagement_events import (
 )
 from analytics_data_api.v0.tests.utils import create_engagement, create_enterprise_user
 from analytics_data_api.v0.tests.views import CourseSamples
-from analyticsdataserver.tests import TestCaseWithAuthentication
+from analyticsdataserver.tests.utils import TestCaseWithAuthentication
 
 PAGINATED_SAMPLE_DATA = [
     {
