+++ analyticsdataserver/router.py
@@ -1,5 +1,7 @@
 from django.conf import settings
 
+from analytics_data_api.middleware import thread_data
+
 
 class AnalyticsApiRouter:
     def db_for_read(self, model, **hints):  # pylint: disable=unused-argument
@@ -7,9 +9,10 @@ class AnalyticsApiRouter:
         return self._get_database(model._meta.app_label)
 
     def _get_database(self, app_label):
-        if app_label in ('v0', 'enterprise_data'):
+        if app_label in ('v0', 'v1', 'enterprise_data'):
+            if hasattr(thread_data, 'analyticsapi_database'):
+                return getattr(thread_data, 'analyticsapi_database', 'default')
             return getattr(settings, 'ANALYTICS_DATABASE', 'default')
-
         return None
 
     def db_for_write(self, model, **hints):  # pylint: disable=unused-argument
